# Fix performance regression from Antimirov derivative mode