A recent improvement to the DFA mode introduced a switch to Antimirov style derivatives and powerset transitions to slow down the growth of a large state space. The change introduced a new branch into the innermost matching loop, which resulted in an ~1.15X slowdown for the common case where Brzozowski derivatives are used.

This pull request pulls that branch out of the innermost loop into an outer loop that checks the condition only every 1000 steps. Here are measurements from 8 benchmarks in [dotnet/performance](https://github.com/dotnet/performance/blob/main/src/benchmarks/micro/libraries/System.Text.RegularExpressions/Perf.Regex.Common.cs) repo that rely on the DFA matching loop (as opposed to being dominated by the prefix search). The new version largely matches the performance of just commenting the whole branch out.

|Benchmark|Branch in loop|Branch commented out|Branch every 1000 steps|
|-|-|-|-|
|Email_IsMatch|300.52 ns|267.12 ns|240.12 ns|
|Email_IsNotMatch|285.29 ns|266.82 ns|299.55 ns|
|Date_IsMatch|178.33 ns|153.21 ns|154.91 ns|
|Date_IsNotMatch|176.34 ns|154.75 ns|148.81 ns|
|IP_IsMatch|161.14 ns|137.60 ns|132.38 ns|
|IP_IsNotMatch|160.54 ns|129.19 ns|126.93 ns|
|Uri_IsMatch|112.71 ns|101.30 ns|83.15 ns|
|Uri_IsNotMatch|196.87 ns|162.79 ns|152.77 ns|

To avoid copy pasting code, the matching code now uses a technique for zero-cost static polymorphism [described here](https://stackoverflow.com/questions/53785910/avoiding-the-overhead-of-c-sharp-virtual-calls). Brzozowski/Antimirov transitions are implemented as empty structs implementing `ITransition`, and the `Delta` function gets one of these as a type parameter. With constraints like `where Transition: struct, ITransition`, a call like `default(Transition).TakeTransition(...)` should be a static call (I'd guess it ends up just being inlined, but haven't checked the IL).